requirements.txt
Django
Pillow
django-modeltranslation
django-ckeditor
